fix(admin models)
admin models moved to models directory

README updated
