+++ models/admin/adminRequest.go
@@ -1,0 +1,5 @@
+package admin
+
+type UpdateUserToAdminRequest struct {
+	AdminPassword string `json:"admin_password" example:"secret"`
+}
