+++ models/admin/unblockRequest.go
@@ -1,0 +1,5 @@
+package admin
+
+type UnblockUserRequest struct {
+	Username string `json:"username" example:"user2"`
+}
